some physics added - try to branch everything from simulation and stick to that single instance

+++ src/main/java/group17/phase1/Titan/Graphics/Renderer/Point3D.java
@@ -6,10 +6,10 @@ public class Point3D implements CoordinateInterface
 {
     public final static Point3D origin = new Point3D(0, 0, 0);
 
-    public float x, y, z;
-    public float xOffset, yOffset, zOffset;
+    public double x, y, z;
+    public double xOffset, yOffset, zOffset;
 
-    public Point3D(float x, float y, float z)
+    public Point3D(double x, double y, double z)
     {
         this.x = x;
         this.y = y;
@@ -35,17 +35,17 @@ public class Point3D implements CoordinateInterface
         return this.z + this.zOffset;
     }
 
-    public static float euclidDist(Point3D p1, Point3D p2) {
+    public static double euclidDist(Point3D p1, Point3D p2) {
         double x2 = Math.pow(p1.x - p2.x, 2);
         double y2 = Math.pow(p1.y - p2.y, 2);
         double z2 = Math.pow(p1.z - p2.z, 2);
-        return (float) Math.sqrt(x2 + y2 + z2);
+        return Math.sqrt(x2 + y2 + z2);
     }
 
-    public float euclidDist(Point3D p2) {
+    public double euclidDist(Point3D p2) {
         double x2 = Math.pow(this.x - p2.x, 2);
         double y2 = Math.pow(this.y - p2.y, 2);
         double z2 = Math.pow(this.z - p2.z, 2);
-        return (float) Math.sqrt(x2 + y2 + z2);
+        return  Math.sqrt(x2 + y2 + z2);
     }
 }
